fix duplicate example names
