+++ packages/dioxus-tui/examples/all_terminal_events.rs
