+++ packages/dioxus-tui/examples/many_small_edit_stress.rs
