+++ packages/router/examples/simple_routes.rs
